feature: show download status in search results
- Create shared API function for downloaded files in src/api/downloaded.ts
- Refactor Downloaded.tsx to use shared getDownloaded function
- Add download status indicator in search results with green badge
- Display "다운로드됨" badge with download icon for already downloaded videos
- Users can now easily identify which videos are already downloaded to avoid duplicates
